Moved the non-test source files from mh/tests to mh/tests/lib
