+++ tests/thread_test.c
@@ -1,4 +1,4 @@
-#include "default_tests.h"
+#include "lib/default_tests.h"
 #include "../inc/mh_thread.h"
 #include <unistd.h>
